requirements.txt
livekit>=1.0.0
livekit-api>=1.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.16.0
aiohttp>=3.9.0 